Rename move_generation module
Renamed move_generation module to move_list.
